Allow speed_scale to be floating point
Jedi Academy for example uses floating point values as font scales
rather than integer font types.

+++ strafe_helper.h
@@ -11,7 +11,7 @@ struct StrafeHelperParams {
     float scale;
     float height;
     float y;
-    int speed_scale;
+    float speed_scale;
     float speed_y;
 };
